new: changed Info View from Component to Container, so that there can contain further components
The XTemplate in the File patient/Info.js  is loaded and populated with data from the selected patient record. The corresponding patientInfo View is then rendered using the XTemplate and the provided Data. The Xtemplate has to be modified for the correct displaying of the Patient Data.  Plz look at the console for 'Object' and expand the console.log item to have a look with data is available in the selected record.